+++ task_files/go_scripts/flutter_mobile_create_layout/template/template_page.dart
@@ -4,6 +4,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'template_riverpod_provider.dart';
+import 'package:eneobia/util/riverpod_providers/wml_colors/wml_colors.dart';
+import 'package:eneobia/util/riverpod_providers/wml_fonts/wml_fonts.dart';
+import 'package:eneobia/util/riverpod_providers/wml_nav/wml_nav.dart';
+import 'package:eneobia/util/riverpod_providers/wml_spacing/wml_spacing.dart';
 
 class WMLTemplateWidget extends ConsumerStatefulWidget {
   WMLTemplateWidget({super.key,required this.pageWidget});
@@ -19,6 +23,12 @@ class _WMLTemplateState extends ConsumerState<WMLTemplateWidget> {
   Widget build(BuildContext context) {
 
     final rp = ref.watch(WMLTemplateRiverpodProvider);
+    final wmlColors = ref.watch(WMLColorsRiverpodProvider);
+    final wmlSpacing = ref.watch(WMLSpacingRiverpodProvider);
+    final wmlFonts = ref.watch(WMLFontsRiverpodProvider);
+    final wmlNav = ref.watch(WMLNavRiverpodProvider);
+    final contextHeight = MediaQuery.of(context).size.height;
+    final contextWidth = MediaQuery.of(context).size.width;
 
     return SafeArea(
         child:Scaffold(