+++ task_files/go_scripts/flutter_mobile_create_page/template/template_page.dart
@@ -4,6 +4,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'template_page_riverpod_provider.dart';
+import 'package:eneobia/util/riverpod_providers/wml_colors/wml_colors.dart';
+import 'package:eneobia/util/riverpod_providers/wml_fonts/wml_fonts.dart';
+import 'package:eneobia/util/riverpod_providers/wml_nav/wml_nav.dart';
+import 'package:eneobia/util/riverpod_providers/wml_spacing/wml_spacing.dart';
 
 class WMLTemplateWidget extends ConsumerStatefulWidget {
   WMLTemplateWidget({super.key});
@@ -16,6 +20,15 @@ class WMLTemplateWidget extends ConsumerStatefulWidget {
 class _WMLTemplateState extends ConsumerState<WMLTemplateWidget> {
   @override
   Widget build(BuildContext context) {
+
+    final rp = ref.watch(WMLTemplateRiverpodProvider);
+    final wmlColors = ref.watch(WMLColorsRiverpodProvider);
+    final wmlSpacing = ref.watch(WMLSpacingRiverpodProvider);
+    final wmlFonts = ref.watch(WMLFontsRiverpodProvider);
+    final wmlNav = ref.watch(WMLNavRiverpodProvider);
+    final contextHeight = MediaQuery.of(context).size.height;
+    final contextWidth = MediaQuery.of(context).size.width;
+
 
     return SizedBox(
         width: MediaQuery.of(context).size.width, child: Text("Hello Word"));